# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/resumes": {
      "get": {
        "operationId": "ResumesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список резюме",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumeResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить все резюме текущего пользователя",
        "tags": [
          "resumes"
        ]
      },
      "post": {
        "operationId": "ResumesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResumeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Резюме успешно создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Slug уже занят"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать новое резюме",
        "tags": [
          "resumes"
        ]
      }
    },
    "/api/resumes/{id}": {
      "get": {
        "operationId": "ResumesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID резюме",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о резюме",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Резюме не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить резюме по ID",
        "tags": [
          "resumes"
        ]
      },
      "put": {
        "operationId": "ResumesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID резюме",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResumeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Резюме обновлено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Резюме не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить резюме",
        "tags": [
          "resumes"
        ]
      },
      "delete": {
        "operationId": "ResumesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID резюме",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Резюме удалено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Резюме не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить резюме",
        "tags": [
          "resumes"
        ]
      }
    },
    "/api/resumes/{id}/versions": {
      "post": {
        "operationId": "ResumesController_createVersion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID резюме",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResumeVersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Версия резюме создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeVersionResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Резюме не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать новую версию резюме",
        "tags": [
          "resumes"
        ]
      }
    },
    "/api/resumes/{id}/publish": {
      "post": {
        "operationId": "ResumesController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID резюме",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Резюме опубликовано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Резюме не найдено"
          },
          "409": {
            "description": "Невозможно опубликовать резюме без версий"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Опубликовать резюме",
        "tags": [
          "resumes"
        ]
      }
    },
    "/api/resumes/{id}/export/pdf": {
      "get": {
        "operationId": "ResumesController_exportPdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID резюме",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF файл резюме",
            "content": {
              "application/pdf": {}
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Резюме не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Экспортировать резюме в PDF",
        "tags": [
          "resumes"
        ]
      }
    },
    "/api/resumes/public/{slug}": {
      "get": {
        "operationId": "ResumesController_getPublicResume",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Slug резюме",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Публичное резюме",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicResumeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Резюме не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить публичное резюме по slug",
        "tags": [
          "resumes"
        ]
      }
    },
...
    "/api/storage/presigned-upload": {
      "post": {
        "operationId": "StorageController_generatePresignedUploadUrl",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Presigned URL для загрузки сгенерирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUploadResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить presigned URL для загрузки файла",
        "tags": [
          "storage"
        ]
      }
    },
    "/api/storage/presigned-download": {
      "post": {
        "operationId": "StorageController_generatePresignedDownloadUrl",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Presigned URL для скачивания сгенерирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedDownloadResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить presigned URL для скачивания файла",
        "tags": [
          "storage"
        ]
      }
    },
    "/api/storage/upload": {
      "post": {
        "operationId": "StorageController_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Файл для загрузки",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Файл успешно загружен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadFileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Загрузить файл напрямую в хранилище",
        "tags": [
          "storage"
        ]
      }
    }
  },
  "info": {
    "title": "PortfolioX API",
    "description": "Backend API для платформы PortfolioX - совместное резюме + портфолио",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ResumeVersionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID версии резюме",
            "example": "clx1234567890"
          },
          "resumeId": {
            "type": "string",
            "description": "ID резюме",
            "example": "clx1234567890"
          },
          "content": {
            "type": "string",
            "description": "Содержимое резюме",
            "example": "{\"sections\": [...]}"
          },
          "versionTag": {
            "type": "object",
            "description": "Тег версии",
            "example": "v1.0",
            "nullable": true
          },
          "aiScore": {
            "type": "object",
            "description": "AI оценка",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "resumeId",
          "content",
          "createdAt"
        ]
      },
      "ResumeCountsDto": {
        "type": "object",
        "properties": {
          "versions": {
            "type": "number",
            "description": "Количество версий",
            "example": 3
          }
        },
        "required": [
          "versions"
        ]
      },
      "ResumeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID резюме",
            "example": "clx1234567890"
          },
          "userId": {
            "type": "string",
            "description": "ID пользователя",
            "example": "clx1234567890"
          },
          "title": {
            "type": "string",
            "description": "Название резюме",
            "example": "Software Engineer Resume"
          },
          "slug": {
            "type": "string",
            "description": "Slug резюме",
            "example": "software-engineer-resume"
          },
          "description": {
            "type": "object",
            "description": "Описание резюме",
            "nullable": true
          },
          "isPublished": {
            "type": "boolean",
            "description": "Опубликовано ли резюме",
            "example": false
          },
          "versions": {
            "description": "Версии резюме",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResumeVersionResponseDto"
            }
          },
          "_count": {
            "description": "Количество версий",
            "allOf": [
              {
                "$ref": "#/components/schemas/ResumeCountsDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата обновления",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "slug",
          "isPublished",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateResumeDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Название резюме",
            "example": "Frontend Developer Resume"
          },
          "slug": {
            "type": "string",
            "description": "Slug резюме (автоматически генерируется из названия, если не указан)",
            "example": "frontend-developer-resume"
          },
          "description": {
            "type": "string",
            "description": "Описание резюме",
            "example": "Резюме для позиции Frontend Developer"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateResumeDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Название резюме",
            "example": "Frontend Developer Resume"
          },
          "description": {
            "type": "string",
            "description": "Описание резюме",
            "example": "Резюме для позиции Frontend Developer"
          },
          "isPublished": {
            "type": "boolean",
            "description": "Статус публикации резюме",
            "example": true
          }
        }
      },
      "CreateResumeVersionDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Содержимое резюме (markdown или JSON)",
            "example": "# Experience\n\n## Company Name\n- Developed..."
          },
          "versionTag": {
            "type": "string",
            "description": "Тег версии",
            "example": "v1.0"
          }
        },
        "required": [
          "content"
        ]
      },
      "PublicResumeUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID пользователя",
            "example": "clx1234567890"
          },
          "name": {
            "type": "object",
            "description": "Имя пользователя",
            "nullable": true
          },
          "headline": {
            "type": "object",
            "description": "Заголовок профиля",
            "nullable": true
          },
          "avatarUrl": {
            "type": "object",
            "description": "URL аватара",
            "nullable": true
          },
          "publicHandle": {
            "type": "object",
            "description": "Публичный handle",
            "nullable": true
          }
        },
        "required": [
          "id"
        ]
      },
      "PublicResumeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID резюме",
            "example": "clx1234567890"
          },
          "userId": {
            "type": "string",
            "description": "ID пользователя",
            "example": "clx1234567890"
          },
          "title": {
            "type": "string",
            "description": "Название резюме",
            "example": "Software Engineer Resume"
          },
          "slug": {
            "type": "string",
            "description": "Slug резюме",
            "example": "software-engineer-resume"
          },
          "description": {
            "type": "object",
            "description": "Описание резюме",
            "nullable": true
          },
          "isPublished": {
            "type": "boolean",
            "description": "Опубликовано ли резюме",
            "example": true
          },
          "user": {
            "description": "Данные пользователя",
            "allOf": [
              {
                "$ref": "#/components/schemas/PublicResumeUserDto"
              }
            ]
          },
          "versions": {
            "description": "Версии резюме",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResumeVersionResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата обновления",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "slug",
          "isPublished",
          "user",
          "versions",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "PresignedUploadResponseDto": {
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string",
            "description": "Presigned URL для загрузки файла",
            "example": "https://s3.example.com/bucket/file?X-Amz-Algorithm=..."
          },
          "fileKey": {
            "type": "string",
            "description": "Ключ файла в хранилище",
            "example": "uploads/123e4567-e89b-12d3-a456-426614174000.jpg"
          },
          "publicUrl": {
            "type": "string",
            "description": "Публичный URL файла",
            "example": "https://storage.example.com/bucket/uploads/123e4567-e89b-12d3-a456-426614174000.jpg"
          }
        },
        "required": [
          "uploadUrl",
          "fileKey",
          "publicUrl"
        ]
      },
      "PresignedDownloadResponseDto": {
        "type": "object",
        "properties": {
          "downloadUrl": {
            "type": "string",
            "description": "Presigned URL для скачивания файла",
            "example": "https://s3.example.com/bucket/file?X-Amz-Algorithm=..."
          }
        },
        "required": [
          "downloadUrl"
        ]
      },
      "UploadFileDto": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "Файл для загрузки"
          },
          "folder": {
            "type": "string",
            "description": "Папка для сохранения файла (опционально)",
            "example": "uploads"
          }
        },
        "required": [
          "file"
        ]
      },
      "UploadFileResponseDto": {
        "type": "object",
        "properties": {
          "fileKey": {
            "type": "string",
            "description": "Ключ файла в хранилище",
            "example": "uploads/123e4567-e89b-12d3-a456-426614174000.jpg"
          },
          "publicUrl": {
            "type": "string",
            "description": "Публичный URL файла",
            "example": "https://storage.example.com/bucket/uploads/123e4567-e89b-12d3-a456-426614174000.jpg"
          }
        },
        "required": [
          "fileKey",
          "publicUrl"
        ]
      }
    }
  }
}